v1.0.0
* Update .gitignore

Updated to use case

* Create package.json

* Current Logic

Current logic:
- It can collect registered Strapi routes to in the future verify if they have a valid `config.body` using Zod
- Initial Zod logic to parse the `config.body` objects
- Middleware that will search inside strapi all routes and verify what route is the current one. So in the future its possible to relate this searched route with all parsed routes and then verify the body

* Zod Error parser

Created error parser to when Zod throws invalid schema.
Parsing the Zod Error saves into the details of the  custom error class all the properties that have problems allowing the end user to easily see where is the problem

* Saved parsed routes into plugin config

Collected parsed routes and saved them into the plugin configuration

* Fix file management to export middlewares

Moved logic at middlewares/index.ts into its dedicated file to fix the middleware exportation not working

* Fixed middleware call

Added missing config property that is sent when the middleware is called

* Middleware order check

Added logic that verifies if "our" middleware is registered in the "config/middlewares" as a middleware that runs after the execution of the Strapi middleware responsible of parsing the request body (strapi::body)

* .

* Yup Validation

Using Yup package from Strapi verifies if the body sent in the request is following the schema defined in the route

* Yup validation error handling

Added catch to when yup validate() method throws error and used Strapi Yup error to throw to the end user a 400 BadRequest

* Refactoring bootstrap

- Moved some methods from utilities into their only file that are using them
- Major refactoring on the config schemas, now each route can have 2 schemas, one for the body and one for the files (this because handling files was being a pain in the ass)
- Changed bootstrap to add this new schemas into a object and the key of the entry is the route method and url, this will make the search for the route schemas more direct
- Commented almost all middleware code because of refactoring, middleware refactor will  be in next commits

* Refactoring middleware

- Refactor middleware because of new changes. Now it also verifies if file is indeed a file
- Files array still need fixing

* Major Refactoring and added array object

- Full refactoring of the code
- For now removed files
- Added way to define one property to be a array of defined objects
- Removed yup middleware validation, now it fully used Zod (validating schema definition and end user schema follow)

* Files schema definition and validation

Added logic that validates the definition of the files schema and that validates the data sent by the end user

* .

* .

* TODO next versions

* Added boolean option

Added way to define boolean in the schema and a boolean parser

* Removed middleware export

Removed export of middlewares to outside of the plugin, now the use of the middlewares is being made by the plugin itself in the bootstrap.
Dont know in where in the list the middleware is added but in the tests it all worked

* Non GETs & upload endpoint

Since Strapi doesnt parse GET methods remove from the routes find the ability to find GETs
Fixed problem with the route with path === "/((?!uploads/).+)" because the regex of it pretty mutch triggers with all the endpoints, added validation if thats the case to find the first route that is not this "upload" route

* Plugin Rename

* Update README.md